Fix online covariance
