+++ cpp/oneapi/dal/algo/covariance/backend/cpu/kernel_utils.hpp
@@ -1,0 +1,29 @@
+/*******************************************************************************
+* Copyright 2021 Intel Corporation
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*******************************************************************************/
+
+#pragma once
+
+#include "oneapi/dal/algo/covariance/compute_types.hpp"
+
+#include "daal/src/algorithms/covariance/covariance_kernel.h"
+
+namespace oneapi::dal::covariance::backend {
+
+using daal_hyperparameters_t = daal::algorithms::covariance::internal::Hyperparameter;
+
+template <typename Float, typename Task>
+daal_hyperparameters_t convert_parameters(const detail::compute_parameters<Task>& params);
+} // namespace oneapi::dal::covariance::backend
